update: insert sql e conecta bd - Inserido o código SQL na pasta de configuração; - Configurado código para leitura do SQL ao invés de string no DBConfig para maior facilidade na manutenção e visualização; - Acrescentado driver postgres no POM
